added animation hair, but not animating
