+++ public/assets/12/objects/the-hair/main.glb
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:fbd79772fa7f9a0f80e2ef04272b332652de8b557e25972b28b269265be61bfe
-size 554872
+oid sha256:6d36ef810ad51d1254979b885387fd5c5e0df81a03e9c270b35c090ed5b1b91a
+size 592520
